Main: fix mac burn long<>GLint
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@20161 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/darwin/OpenGLTestDarwin.cpp
@@ -20,7 +20,7 @@ bool is3DAccelerationSupported()
     CGDirectDisplayID   display = CGMainDisplayID ();
     CGOpenGLDisplayMask cglDisplayMask = CGDisplayIDToOpenGLDisplayMask (display);
     CGLPixelFormatObj   pixelFormat = NULL;
-    long numPixelFormats = 0;
+    GLint numPixelFormats = 0;
 
     CGLPixelFormatAttribute attribs[] = {
         kCGLPFADisplayMask,
